fix(use-widget-color-state): add `useWidgetColorSet`
