+++ src/services/dashboards/widgets/_base/base-pie/BasePieWidget.vue
@@ -43,7 +43,6 @@ import {
 import type { Location } from 'vue-router/types/router';
 
 import { PDataLoader } from '@spaceone/design-system';
-import dayjs from 'dayjs';
 
 import { getPageStart } from '@cloudforet/core-lib/component-util/pagination';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
@@ -68,6 +67,7 @@ import {
     getPieChartLegends, getRefinedPieChartData,
 } from '@/services/dashboards/widgets/_helpers/widget-chart-helper';
 import { getReferenceTypeOfGroupBy } from '@/services/dashboards/widgets/_helpers/widget-table-helper';
+import { useWidgetColorSet } from '@/services/dashboards/widgets/_hooks/use-widget-color-set';
 import { useWidgetFrameProps } from '@/services/dashboards/widgets/_hooks/use-widget-frame-props';
 import { useWidgetLifecycle } from '@/services/dashboards/widgets/_hooks/use-widget-lifecycle';
 // eslint-disable-next-line import/no-cycle
@@ -87,6 +87,7 @@ const chartHelper = useAmcharts5(chartContext);
 const props = defineProps<WidgetProps>();
 const state = reactive({
     ...toRefs(useWidgetState<FullData>(props)),
+    ...useWidgetColorSet({ theme: computed(() => props.theme), data: computed(() => state.chartData) }),
     chart: null as null|ReturnType<typeof chartHelper.createPieChart | typeof chartHelper.createDonutChart>,
     series: null as null|ReturnType<typeof chartHelper.createPieSeries>,
     chartData: computed<PieChartData[]>(() => {
@@ -105,8 +106,8 @@ const state = reactive({
     legends: [] as Legend[],
     thisPage: 1,
     dateRange: computed<DateRange>(() => ({
-        start: dayjs.utc(state.settings?.date_range?.start).format('YYYY-MM'),
-        end: dayjs.utc(state.settings?.date_range?.end).format('YYYY-MM'),
+        start: state.settings?.date_range?.start,
+        end: state.settings?.date_range?.end,
     })),
     widgetLocation: computed<Location>(() => ({
         name: COST_EXPLORER_ROUTE.COST_ANALYSIS._NAME,
