+++ src/services/dashboards/widgets/_hooks/use-widget-color-set.ts
@@ -1,0 +1,39 @@
+import type { ComputedRef } from 'vue';
+import { computed } from 'vue';
+
+import { palette } from '@/styles/colors';
+
+import type { WidgetColorSetType, WidgetTheme } from '@/services/dashboards/widgets/_configs/view-config';
+import { WIDGET_THEMES } from '@/services/dashboards/widgets/_configs/view-config';
+
+
+const getColorSet = (theme: WidgetTheme, colorSetType: WidgetColorSetType = 'basic'): string[] => {
+    let colorSet = WIDGET_THEMES.map((d) => palette[d][400]);
+    if (colorSetType === 'massive') {
+        const colors1 = WIDGET_THEMES.map((d) => [palette[d][400], palette[d][600]]).flat();
+        const colors2 = WIDGET_THEMES.map((d) => [palette[d][500], palette[d][700]]).flat();
+        colorSet = colors1.concat(colors2);
+    }
+    const themeIndex = WIDGET_THEMES.findIndex((d) => d === theme);
+    if (themeIndex > -1) {
+        const arr1 = colorSet.slice(themeIndex, colorSet.length);
+        const arr2 = colorSet.slice(0, themeIndex);
+        return arr1.concat(arr2);
+    }
+    return colorSet;
+};
+
+interface UseWidgetColorSetOptions {
+    theme: ComputedRef<WidgetTheme | undefined>;
+    data: ComputedRef<any[] | undefined>;
+}
+export const useWidgetColorSet = ({
+    theme, data,
+}: UseWidgetColorSetOptions) => ({
+    colorSet: computed<string[]>(() => {
+        if (!theme?.value) return [];
+        if (!data || !Array.isArray(data.value)) return [];
+        const colorSetType: WidgetColorSetType = data.value.length > 9 ? 'massive' : 'basic';
+        return getColorSet(theme.value, colorSetType);
+    }),
+});
